requirements.txt
Django==5.2.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
Pillow==10.4.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
